requirements.txt
streamlit
ultralytics
opencv-python
wget
pillow
torch